+++ modules/azure-backup-vault/_examples/blobs/example.tf
@@ -10,6 +10,7 @@ module "azure_backup_vault" {
   backup_resource_group_name = "example-backup-rg"
   vault = {
     name           = "example-backup-vault"
+    location       = "eastus"
     datastore_type = "VaultStore"
     redundancy     = "LocallyRedundant"
   }
@@ -26,6 +27,7 @@ module "azure_backup_vault" {
   blob_instances = [
     {
       name = "blob1-backup"
+      location = "eastus"
       storage_account_name = "example-storage-account"
       storage_account_resource_group = "example-storage-rg"
       storage_account_container_names = ["container1"]