+++ modules/azure-backup-vault/variables.tf
@@ -22,6 +22,7 @@ variable "vault" {
   description = "Backup vault configuration"
   type = object({
     name                         = string
+    location                     = string
     datastore_type               = string
     redundancy                   = string
     cross_region_restore_enabled = optional(bool)
@@ -60,6 +61,7 @@ variable "disk_instances" {
   description = "List of backup instances for disks"
   type = list(object({
     name                = string
+    location            = string
     disk_resource_group = string
     policy_key          = string
   }))
@@ -100,6 +102,7 @@ variable "blob_instances" {
   description = "List of backup instances for blobs"
   type = list(object({
     name                            = string
+    location                        = string
     storage_account_name            = string
     storage_account_resource_group  = string
     storage_account_container_names = optional(list(string))
@@ -145,6 +148,7 @@ variable "postgresql_instances" {
   description = "List of backup instances for PostgreSQL Flexible Server"
   type = list(object({
     name                = string
+    location            = string
     server_name         = string
     policy_key          = string
     resource_group_name = string
@@ -189,6 +193,7 @@ variable "mysql_instances" {
   description = "List of MySQL Flexible Server backup instances"
   type = list(object({
     name                = string
+    location            = string
     server_name         = string
     resource_group_name = string
     policy_key          = string
@@ -233,6 +238,7 @@ variable "kubernetes_instances" {
   description = "List of Kubernetes cluster backup instances"
   type = list(object({
     name                         = string
+    location                     = string
     cluster_name                 = string
     resource_group_name          = string
     snapshot_resource_group_name = string